[#240]refactor: 마이페이지에 공통 PhoneInput 컴포넌트 적용
