+++ client/src/components/my/MyPagePhoneInput/index.spec.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { render } from '@testing-library/react';
 import 'jest-styled-components';
