+++ client/src/components/my/MyPagePhoneInput/index.style.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const MyPagePhoneInputWrap = styled.div`
+  width: 370px;
+  box-sizing: border-box;
+  font-size: 14px;
+`;
+
+export default { MyPagePhoneInputWrap };
